fix(ios): gradeValue Int? → Double? — supports decimal BGS/CSG grades
Drew's diagnostic question surfaced two hard bugs in the initial
113f091 iOS edit:

Bug A — write path silently drops fractional grade:
  Int("9.5") returns nil. mapCardItemToHolding produced
  canonicalValue=nil for any BGS 9.5 / CSG 8.5 holding. Backend
  received gradeCompany="BGS" + gradeValue=null → treats as ungraded →
  falls into raw bucket → wrong-grade pricing (silent, no error).

Bug B — read path crashes JSONDecoder on decimal:
  Swift's JSONDecoder is strict: decoding a JSON number 9.5 into Int?
  throws DecodingError.dataCorrupted with "Parsed JSON number 9.5
  does not fit in Int". Hard-fails the entire holding decode — readPath
  aborts on first BGS 9.5 holding, breaking sync for affected users.

Backend backfill (8b4465c) cleanly writes gradeValue: 9.5 (Double)
into Cosmos for any BGS/CSG decimal-grade label. The bug would
manifest the next time iOS readPath ran for such a user.

Fix: gradeValue: Int? → Double?

  - InventoryCard.gradeValue: Int? → Double?
  - CompIQResolvedVariant.gradeValue: Int? → Double?
  - PortfolioAddFlowView: Int(viewModel.gradeValue) → Double(...)
  - PortfolioSyncService.mapCardItemToHolding: Int(card.grade) →
    Double(card.grade)
  - PortfolioSyncService.mapHoldingToNewCardItem + updateCardItem:
    holding.gradeValue.map(String.init) → use formatGradeValue helper
    that renders integer-valued doubles as "10" (not "10.0") matching
    the legacy CardItem.grade String format.

Helper: PortfolioSyncService.formatGradeValue(_:) — renders Double as
String with no spurious .0 suffix for integer values. Handles non-finite
defensively (returns "" for NaN/infinity though both should be
impossible from real grade data).

Why Double (not String):
  - Swift JSONDecoder is strict — backend writes JSON number; flipping
    iOS to String would require backend changes OR a custom decoder.
  - Backend CompIQEstimateRequest type contract is `gradeValue: number`
    — Double matches.
  - Cardsight translator does String(...).trim() defensive coercion —
    no symmetric change needed.

Tests added to PortfolioSyncMapperTests.swift (6 new):
  - testMapCardItemToHoldingSendsCanonicalPSA10 — integer round-trip
  - testMapCardItemToHoldingSendsCanonicalDecimalBGS95 — 9.5 round-trip
  - testMapCardItemToHoldingSendsNilCanonicalForRawCards
  - testMapHoldingToNewCardItemDecodesDecimalGradeValueWithoutCrashing
    — Bug B regression guard
  - testMapHoldingToNewCardItemRendersIntegerDoubleWithoutDotZero
    — display parity guard ("10" not "10.0")
  - testFormatGradeValueIntegerProducesNoDecimal +
    testFormatGradeValueDecimalPreserved — helper unit tests

UNVERIFIED on Mac compile. Drew runs xcodebuild + test target on Mac
before merging this branch into main.

References:
  - 113f091 (initial iOS edit — this commit fixes the Int? bug Drew
    surfaced)
  - 8b4465c (backend grade-parser + backfill — writes Double 9.5 to
    Cosmos, which Bug B would crash on)
  - cardsight.translator.ts:33,61-62 (backend coerces number→string)

+++ HobbyIQ/CompatibilityShims.swift
@@ -1388,8 +1388,11 @@ struct CompIQResolvedVariant: Identifiable, Hashable, Codable {
     // when the user has selected a graded slab; downstream
     // addInventoryCard call site passes them through to InventoryCard
     // so the backend's autoPriceHolding read contract finds them.
+    //
+    // gradeValue is Double (not Int) so decimal grades like BGS 9.5 /
+    // CSG 8.5 don't truncate or fail JSONDecode on the wire.
     let gradeCompany: String?
-    let gradeValue: Int?
+    let gradeValue: Double?
     let serialNumber: Int?
     let isAuto: Bool
 
@@ -1403,7 +1406,7 @@ struct CompIQResolvedVariant: Identifiable, Hashable, Codable {
         parallel: String? = nil,
         grade: String? = nil,
         gradeCompany: String? = nil,
-        gradeValue: Int? = nil,
+        gradeValue: Double? = nil,
         serialNumber: Int? = nil,
         isAuto: Bool = true
     ) {
